Allow Asterisk to compile under GCC 4.10
This resolves a large number of compiler warnings from GCC 4.10 which
cause the build to fail under dev mode. The vast majority are
signed/unsigned mismatches in printf-style format strings.
........

Merged revisions 413586 from http://svn.asterisk.org/svn/asterisk/branches/1.8
........

Merged revisions 413587 from http://svn.asterisk.org/svn/asterisk/branches/11
........

Merged revisions 413588 from http://svn.asterisk.org/svn/asterisk/branches/12


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@413589 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_festival.c
@@ -424,7 +424,7 @@ static int festival_exec(struct ast_channel *chan, const char *vdata)
 	/* Convert to HEX and look if there is any matching file in the cache 
 		directory */
 	for (i = 0; i < 16; i++) {
-		snprintf(koko, sizeof(koko), "%X", MD5Res[i]);
+		snprintf(koko, sizeof(koko), "%X", (unsigned)MD5Res[i]);
 		strncat(MD5Hex, koko, sizeof(MD5Hex) - strlen(MD5Hex) - 1);
 	}
 	readcache = 0;
