+++ apps/app_queue.c
@@ -2355,12 +2355,12 @@ static void device_state_cb(void *unused, struct stasis_subscription *sub, struc
 	ao2_iterator_destroy(&qiter);
 
 	if (found) {
-		ast_debug(1, "Device '%s' changed to state '%d' (%s)\n",
+		ast_debug(1, "Device '%s' changed to state '%u' (%s)\n",
 			dev_state->device,
 			dev_state->state,
 			ast_devstate2str(dev_state->state));
 	} else {
-		ast_debug(3, "Device '%s' changed to state '%d' (%s) but we don't care because they're not a member of any queue.\n",
+		ast_debug(3, "Device '%s' changed to state '%u' (%s) but we don't care because they're not a member of any queue.\n",
 			dev_state->device,
 			dev_state->state,
 			ast_devstate2str(dev_state->state));
@@ -6612,7 +6612,7 @@ static int try_calling(struct queue_ent *qe, struct ast_flags opts, char **opt_a
 					ast_monitor_start(which, qe->parent->monfmt, ast_channel_uniqueid(qe->chan), 1, X_REC_IN | X_REC_OUT, NULL);
 				} else {
 					/* Last ditch effort -- no channel, make up something */
-					snprintf(tmpid, sizeof(tmpid), "chan-%lx", ast_random());
+					snprintf(tmpid, sizeof(tmpid), "chan-%lx", (unsigned long)ast_random());
 					ast_monitor_start(which, qe->parent->monfmt, tmpid, 1, X_REC_IN | X_REC_OUT, NULL);
 				}
 				if (!ast_strlen_zero(monexec)) {
